open2 is always passed at least empty attributes, assure that mode is set
If we are doing a create, since we always pass at least empty attributes,
we MUST set mode. Also, the implementations were testing for a NULL pointer
to determine if any attributes were provided. Since the pointer is always
non-NULL, the logic needs to change to check if any attributes OTHER than
mode are set.

Change-Id: Id5e5fdaae804c2cdc07ca13e212e9cfede4426e3
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>