Fix file handle bytes switching in xdr_nfs_fh4 XDR_ENCODE
FSAL_PROXY tests reveals a "Stale file handle" error at
first ls on root directory. This was due to call to htons
in the encode way that was switching bytes and corrupting
the file handle in FSAL_PROXY. Only for XDR_ENCODE case
in xdr_nfs_fh4 function, this patch adds the use of a buffer
to copy and safely switch bytes of nfs_fh4_val without
changing original FSAL_PROXY handle value.

Change-Id: Ia682d0ce83d19ee873c1c7a993e6c86499c48bd7
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>