Re-add owner validation cache optimization
2.3 checked the owner against the cached owner unconditionally
(regardless of cache validity) when doing permission checks for
read/write.  2.4-dev removed this, causing a getattr() call on each
write (since write invalidates attributes).

Re-add this optimization, as it is a 30%-80% performance hit on small
writes, and owner changes infrequently.  In the future, finer grained
cache validity is likely needed.

Change-Id: Ida91b03cb5791bf767161beb7829d51122d0a83b
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>