Fix create operations with GID for non-root users
AIX and Irix NFS clients send create with gid if the parent directory
has setgid bit. This is unnecessary, but ganesha server ends up failing
the request as non-root users can't set uid or gid. Added a workaround
to remove attributes uid and gid for a create request for non-root
users.

Change-Id: Ie6c4599c0bfc476fccf36318ce1733ab4cf5f1a2
Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>