FIX FSAL_PROXY : clear CTIME in setattrs
No way to update CTIME using a NFSv4 SETATTR.
Server will return NFS4ERR_INVAL (22).
time_metadata is a readonly attribute in
NFSv4 and NFSv4.1. (section 5.7 in RFC7530 or RFC5651)

Nevermind : this update is useless, we prevent it
in FSAL_PROXY by clearing CTIME bitmask.

We add a comment in the fsal_api.h to warn people.

Change-Id: I5051992035d0c42d77d67468c0d8f223198bc097
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>