If an exclusive create is replayed, only check owner of file
If an exclusive create is replayed, for example, as a result of the
following system call:

    open("foo", O_RDWR | O_CREAT | O_EXCL, 0);

The normal permission check on replay would fail (NFS v4.1 only since
it requires setting mode on the create).

By skipping permission check if the caller is the owner, we nicely
do the right thing. Allow the replay to succeed because the caller
is the owner (and will have the right verifier).

Change-Id: Id5e9895c0fa672b24cb101cce0ffcb62d37b9729
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>