FSAL_VFS: Handle an odd exclusive create race
It's possible we get down to open2 with a replayed exclusive create
where the upper layers have not found the file, but indeed it exists
(most likely due to some race) so the O_CREAT | OEXCL open fails.

Retry a regular open and then check the verifier and if all pans out,
continue to create the fsal_obj_handle and do the rest of the creation
steps.

Change-Id: Ie886c64f2c7e2164965cc304fecc7d0ebac406cf
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>