FSAL_GLUSTER: Remove glusterfs_fetch_attrs and attributes from obj_handle
glusterfs_fetch_attrs is only used by reopen2 after truncate, but the
caller does not expect attributes to be updated (and the go into the
bit bucket anyway). FSAL_MDCACHE invalidates the attributes in reopen2
if there is a truncate.

This was the last use of the attributes in glusterfs_handle.

Change-Id: I9df4a09c329966e6707a0b7656afd2db650fd712
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>