ceph: pass in correct temporary fd variable to ceph_ll_create
In the UNCHECKED create case, if we get back -EEXIST on the first
ceph_ll_create call, we'll call it again without O_EXCL. The second call
however has the return Fh pointer set to &my_fd->fd, which gets
clobbered just afterward if that call succeeds.

I think this is wrong -- we want to pass in &fd there so that my_fd
gets set up properly.

Change-Id: Ic2167c1664280ea8de2a3dc2ad89769534ce4088
Signed-off-by: Jeff Layton <jlayton@redhat.com>