Improve atomic create
A user space NFS server has a unique problem in that a create request
(CREATE or OPEN/CREATE) is required to create a file, and set attributes
on the file. Even if the source is a POSIX open() call, an EXCLUSIVE
or EXCLUSIVE_41 create must set the verifier into atime and mtime.

This patch introduces a way for the FSAL to indicate the level of
atomic create and setattrs it provides.

The basic level is no special support, in which case, the fsal_helper.c
function create_helper() will generate a random file name (of the form
nfs_ganesha_XXXXXXXX) and then create the file using that name, only
when everything is complete is the file linked with the requested name.

A second level of support uses O_TMPFILE to create an unnamed file
which has the advantage over the random file name in that if Ganesha
crashes, the temporary file will disappear.

The final level of support is for FSALs that can completely support
an atomic create and setattrs.

For reasons of avoiding regressions, and due to the special nature
of the FSAL, FSAL_RGW claims complete atomic create support.

Change-Id: I666771af7193b92e701c50c4346009613bf60692
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>