build: don't use __DATE__ and __TIME__ for Release builds
Current Debian-based distributions such as Debian Stretch,
Ubuntu Xenial, and Ubuntu Yakkety, and soon Fedora, build
with -Werror and -Wdate-time. Such builds fail unless the
places where __DATE__ and __TIME__ are patched out, which
is a nuisance to have to do when building community packages
for download.

Also the #define VERSION line in config.h was not being
substituted correctly.

Change-Id: I8762ebcd921413fe4cb6852946baa4f3a95a3f4d
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>