Remove NOENT from mdcache_readdir and _9p_readdir
The current code in mdcache_readdir() sets status to FSAL status
ERR_FSAL_NOENT when EOD is reached. Removed the return code as
NFS3ERR_NOENT is not listed in the ERRORS list of
READDIRPLUS3 Procedure in the RFC. Also the RFC doesn't clearly mention
that an error should be returned when EOD is reached.
Removed the NOENT check in _9p_readdir() as well.

Change-Id: I13d91478823e6a758f3ad4d537430490718c41d8
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>