MDCACHE - don't cache large directories
Historically, Ganesha has cached the entire set of dirents for a
directory.  This can cause problems, as the directory is locked while
the dirent cache is filled, and no other operations can occur.  In
addition, it can use unbounded memory for the dirent cache of large
directories.

To fix this, add a new configurable, Dir_Max, that indicates the largest
size directory that will be cached.  If a directory larger than this is
encountered, it is marked as uncached, and no dirents for that directory
are cached.

Change-Id: I896b71ce44f9c8fccc493f99af0640aedfa76f97
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>