ceph: workaround for bad mount path handling in libcephfs
libcephfs has had a long standing problem in that it did not set the
path to be mounted properly from the ceph_mount codepath, leading it to
always do an open session request from root with most userland programs.
This was never a problem with ceph-fuse, as it was handled there via the
"client_mountpoint" config option.

The fix is simple and should eventually trickle out to stable ceph
releases, but for now we can work around this by setting the path in
the configuration before we issue the mount request. This workaround
should be no problem with later ceph releases as well.

Change-Id: I18b6de0017673d74bf97b2948e77a7a979d2110b
Signed-off-by: Jeff Layton <jlayton@redhat.com>