Fix use-after-free during shutdown
The root object for an export was having it's sentinal ref released by
the export code; however, this code doesn't own that ref, the hash table
in MDCACHE owns that ref, so when the objest is freed, the hash removes
it from the table after it's freed.

Instead, pass the root object into the FSAL's unexport() op, allowing
MDCACHE to clean up it's state on that object before the last ref is
dropped.

Change-Id: I768306e22cfb3a9b6344c76d0e6a17d59ee29769
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>