display buffer overflow: Fix off by one
Address SANitizer (ASAN) blows up on write right after the end of the
buffer.

I *think* the function should look behind for complete utf8 chars
(instead of looking at current position) because it's always called
with exactly enough room for '...' + terminating null byte.
Even when we force the overflow, it's called with b_current set after
the last character, so the function probably used to print one byte of
garbage in this case.

Change-Id: I8df14a2f51824e466600916a89bb54ab04b14635
Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>