FSAL_RGW: fix mount failure of nfs v3 on windows
When we tried to mount rgwfs on windows using nfsv3, it failed with
a not found errcode 53:
	> mount \\<IP>\mybucket x:\
It is because win client pass a "/mybucket" for export dir, but
ganesha is exporting the desired dir as "mybucket"(without slash)
	# showmount -e
	Export list for <myhost>:
	mybucket               (everyone)
so ganesha could not serve.

Here we export rgwfs with config "Path = mybucket".
We shall fail to export if provided "Path = /mybucket", because
FSAL_RGW directly pass "/mybucket" to librgw as parameter 'path',
which will be digested as a bucket name by librgw.

Actually librgw would like a relative path as a bucket name, then
we could export with config "Path = /mybucket"(with slash) which
will make nfs client succeed to mount on windows.

Change-Id: I02c241b12d77a747e8b039c5ac3a0662351011c9
Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>