release file sys after file sys is unclaimed
Currently in FSAL_GPFS after last export of a file system is removed,
the file system is unclaimed but not removed. If a file system is
deleted and a new file system is added which has the same
dev major, minor numbers as the previous file system then it is
skipped as duplicate file system.
Made changes by calling "release_posix_file_system_with_fslock" after
the file system is unclaimed.

Change-Id: I800d509b220c9300fb70a62b1399fa7482c6f32a
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>