WIP Add directory chunking for readdir.
Added directory invalidate on create and rename into or within
directory.

Unlink and rename out of directory just marks entry deleted so it will
not enumerate from chunk cache with readdir.

Note that mdc_lookup_uncached no longer invalidates the attributes of
the parent directory. I think that was wrong in the first place (we
don't if an open by name "discovers" a new file that we weren't
responsible for creating). This allows the invalidate flag for
mdcache_alloc_and_check_handle to also be used to determine if we need
to invalidate the chunk cache.

Change-Id: Id4aa058bb5e96c26717234c42a4aa761bd80db67
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>