Reduce and mitigate a rename/readir race window
There is a race window in rename() with readdir().  This steps are this:
mdcache_rename() calls the sub-FSAL rename(), which does the actual
rename.  Then, it locks the src and dst dirs, and renames the dirent.
However, between the real rename and the dirent rename, another client
may have come in and done a readdir(), refreshing the dirent cache.  The
lookup on the old dirent then fails, returning ENOENT to  the client.

Reduce this race window, by moving the locking to immediately after the
rename.  This should also make it much less likely that the dirents are
refreshed, since we haven't marked them stale yet.

In addition, handle ENOENT on the src dirent as an indication that this
race has occured, and do not treat it as a user-visible error.

RH Bug #1428798

Change-Id: I9a7e706b25a8a2f1df54a982b2952aad7134a89d
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>