gpfs: enable large ACLs
Currently there is a fixed size buffer of 4K allocated in a struct.
GPFSFSAL_setattrs() doesn't validate and ends up corrupting memory if
more than 4K is needed.  GPFSFSAL_getattrs() ends up returning an
incorrect error ENOSPC.

GPFSFSAL_setattrs knowns the size in advance, so it now uses alloca if
the size needed is more than what is in the struct. GPFSFSAL_getattrs
has a retry mechanism if the passed in buffer is smaller than a required
buffer.

Change-Id: I72528f841966a728e0dfff5ba88761e77f5a8e14
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>