FSAL_RGW: fix rename causing mdcache to have conflict dentries
FSAL_RGW use {bucket,object} pair as a key for mdcache dentries,
so when we rename a file in FSAL_RGW, we change the actual key
inevitably.
But mdcache tries to do a mdcache_key_dup() from
old dentry to new dentry on rename, so we have a new dentry with
old {bucket,object} pair as key.
Upon next lookup for the new file, we should get EEXIST from mdcache.

To fix it, let FSAL_RGW's rename return a special code to signal
mdcache to drop the old dentry only without adding a new one.

Change-Id: I2c25748424f89cd4781805b193a40032c654cefd
Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>