MDCACHE - Handle FSALs where rename doesn't preserve key
There are FSALs (currently only FSAL_RGW) where a rename of an object,
even within the same directory, changes the key used by MDCACHE.  This
means that a renamed dirent will point to the wrong object.

To accomplish this, add a new fso flag indicating that a FSAL keeps key
on rename.  It's true on all FSALs other than RGW at this point.

Change-Id: I65bd2b1c9ad857fbc7ef5cde2c3be1b1ca4dbadf
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>