rpm: handle errors when SIGHUP'ing dbus
Prior to this change, if the nfs-ganesha RPM was installed on a host
without a dbus-daemon process running, the RPM transaction would print the
following errors:

  dbus-daemon: no process found
  warning: %post(nfs-ganesha-2.5-0.dev.17.el7.x86_64) scriptlet failed, exit status 1
  Non-fatal POSTIN scriptlet failure in rpm package nfs-ganesha-2.5-0.dev.17.el7.x86_64

For example, this can happen if yum installs nfs-ganesha in a CentOS
Docker container where killall cannot find a dbus-daemon process.

Flip the ordering of STDIN and STDOUT redirection so that both go to
/dev/null. This properly silences the "dbus-daemon: no process found"
message. This appears to be the original intention of commit
3abb7d3e21769966da320efd40a0837527a5346b that introduced the SIGHUP
step.

Append "|| :" to the command so that the overall exit code will still be
zero even when the killall command fails with a non-zero exit code.
This change will remove the "warning:" message during the %post
scriptlet.

Change-Id: Ie28c4d249bb644cbae18a97bf70df66b44225c46
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>