WIP FSAL_PROXY (from v4 to v4.1) : slotid, sequenceid, sessionid
This patch is the first of a bunch dedicated to migrate the FSAL_PROXY
from NFSv4 to NFSv4.1 protocol to connect to the background server.

First, we change "minorversion" and add slotid and sequenceid in rpc
context. Then we update pxy_clientid_renewer to add CREATE_SESSION and
replace by EXCHANGE_ID.

We add SEQUENCE op.

We further need to move each op from 4 to 4.1. But before doing that,
we post this patch as WIP to get back comments.

Change-Id: I30ea8fe10a0649c84fecb3cc094dc16a62a5a65a
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>