[FSAL_GPFS] Fix race between setting/consuming FD on export.
On export creation, the GPFS fsal_up thread requires the file-descriptor
to be set before being able to update the inode with callback information.
Especially on POWER systems it does happen that the main-thread,
which is setting the FD, is only able to do this after the fsal_up thread
has tried to do this a few times with an unset(invalid) FD.
Therefore, the FD is now set within the fsal_up thread before
triggering the inode update.

Change-Id: I847af6e0642c91d4e3cd87b6b74bb3c68c8e039b
Signed-off-by: Swen Schillig <swen@vnet.ibm.com>