MDCACHE - Ensure parent lookups use wire handle
In order for create_handle() to work, it needs to be passed a wire
handle, not a key.  However, the parent pointer was a key.  This worked
for most FSALs, since they used the entire handle as the key, but a few
FSALs, notably GPFS, only use a subset of the handle as the key, causing
parent lookups to fail and always create a new handle.

To fix this, make the parent pointer be a full handle, rather than a
key.  At some point in the future, it may be advisible to revisit the
handle/key API with the FSALs to streamline this.

Change-Id: I084d46b9441cdce0701da977c479cef34cfd0f52
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>