readdir: Allow for FSALs that can't use a cookie for whence
For some FSALs it's impractical to resume a directory listing using a
cookie. These FSALs however can use the name of the previous dirent.

For these FSALs, for the whence value, we either pass NULL to indicate
start from beginning, or the name (cast to fsal_cookie_t * - the FSAL
will need to re-cast to char *).

If we get a non-zero whence from the protocol layer and we don't have
it cached, we will re-scan the directory from the beginning. We do
skip re-scanning should we have any chunks cached.

Change-Id: I337eefc879a83405be4bc24b7658ee6380d394b3
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>