Fixup handle/key usage
During the MDCACHE conversion, the usage of wire handles and keys was
conflated.  This didn't cause problems for most FSALs, since they used
the same thing for both, but it was broken none-the-less.

Clean this up, by removing the calls to extract_handle() before calls to
create_handle().  This means that create_handle() always takes a wire
handle, not a key.

Change MDCACHE to handle this, by extracting the key for lookups, but
using the wire handle for handle creation.

Extend the endianess flags to the create_handle() call, so that it can
be passed into MDCACHE.

Change-Id: I67d19ef341bcefcf4b76c45a073eaa5778e3a266
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>