Fix race between mdcache_unexport and mdc_check_mapping
Once we start an unexport, we must reject any new mappings between the
export and cache entries. In fact, we can utilize mdc_check_mapping to
shortcut and return ERR_FSAL_STALE for lookups and other operations
that access an mdcache entry (they come through mdc_check_mapping to
assure that the entry is mapped for an export it is accessed via).

Because of the way the locking works, we check for unexport again
before actually creating a new mapping just in case we had an unusual
race between mdcache_unexport and mdc_check_mapping.

Change-Id: I68b600c1996c7f41c7cd0b3124ce8cf1ac5c772a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>