Fix race between mdcache_unexport and mdc_check_mapping
Once we start an unexport, we must reject any new mappings between the
export and cache entries. In fact, we can utilize mdc_check_mapping to
shortcut and return ERR_FSAL_STALE for lookups and other operations
that access an mdcache entry (they come through mdc_check_mapping to
assure that the entry is mapped for an export it is accessed via).

Because of the way the locking works, we check for unexport again
before actually creating a new mapping just in case we had an unusual
race between mdcache_unexport and mdc_check_mapping.

Move the mapping for new entries into mdcache_lru_get so that it is
complete before anyone else can see the entry. Note that once it's
mapped, it is visible to unexport, and then unexport could end up
with the last LRU unref, so we assure that the entry is capable of
being cleaned up on unexport's LRU unref. With this, mdcache_lru_get
can now return NULL if an unexport is in progress (in which case it
just dumps the new entry - no way to check that BEFORE pulling an
entry off of L2 or L1 unfortunately).

Change-Id: I68b600c1996c7f41c7cd0b3124ce8cf1ac5c772a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>