Allow NO-OP SETATTR of owner_group even if when not a member
Linux seems to allow a user to chgrp a file to an owner_group
they are not a member of if the file already belongs to that
owner_group.

Note: This change enables a theoretical race if the FSAL or
the underlying filesystem does not do a permission check:

1. File is owned by user1:group2, user1 is not a member of group2
2. Ganesha caches attributes
3. Admin locally changes ownership of the file to user1:group1
4. Supposedly NO-OP SETATTR comes in to change to user1:group2
   thus overriding the change the admin made since Ganesha's cached
   owner_group for the file is group2 and thus appears to be a
   NO-OP.

We could instead detect the NO-OP and not set the owner_group but
then an admin will come along and complain their change didn't
take place due to Ganesha's stale cache...

Note that Ganesha also allows races where the ownership of the
file has changed, so we really aren't introducing anything
dramatically new.

Change-Id: I8fb50409b67cc46f414f8c7eba4a75debd6f6ec8
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>