build: nfs-ganesha-gluster-selinux sub-package to access to FUSE file systems
nfs-ganesha+glusterfs+pacemaker HA uses a glusterfs (FUSE) shared-storage
volume for sharing state between the ganesha.nfsd that comprise the cluster.

As SELinux continues to evolve we need the ability to set new SELinux
booleans at install time. At this time there is only one SELinux boolean
specific to GlusterFS. In the future if we find we need to set SELinux
booleans that are common to multiple FSALs or to NFS-Ganesha in general we
may introduce an umbrella nfs-ganesha-selinux sub-package; for now we will
make this a gluster specifc

Change-Id: I47c0ab7672c0056bb19a5754b2a82117f52b949d
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>